renta_autos_app/requirements.txt
mysql-connector-python
customtkinter
bcrypt
python-dotenv